sort out and relocate tests
